examples/python/langchain-agent/requirements.txt
capsule-run
langchain-openai
langgraph
python-dotenv
